Use a fixed-size type for the buffer index to avoid 32Bit vs. 64Bit issues

+++ brian2/devices/device.py
@@ -357,9 +357,9 @@ class RuntimeDevice(Device):
         #: last reference to the `Variable` object used as a key is gone
         self.arrays = WeakKeyDictionary()
         self.randn_buffer = np.zeros(1024, dtype=np.double)
-        self.randn_buffer_index = np.zeros(1, dtype=np.int)
+        self.randn_buffer_index = np.zeros(1, dtype=np.int32)
         self.rand_buffer = np.zeros(1024, dtype=np.double)
-        self.rand_buffer_index = np.zeros(1, dtype=np.int)
+        self.rand_buffer_index = np.zeros(1, dtype=np.int32)
 
     def get_array_name(self, var, access_data=True):
         # if no owner is set, this is a temporary object (e.g. the array
